4th exercise
This is getting harder!
