+++ 05-crud-rest-api/exercises/05-04-todo-delete/README.md
@@ -1,0 +1,33 @@
+# Todo DELETE API for deleting todo
+
+## Objective:
+
+- Understand the DELETE API
+
+## Exercise:
+
+- Update Postman to have DELETE method
+- Implement DELETE in router using `deleteTodo` from model
+- Handle error in case of the todo not found.
+
+Request:
+
+```
+DELETE /todos/2
+```
+
+Response:
+
+```json
+{
+  "data": {
+    "id": 2,
+    "title": "Go to school",
+    "description": ""
+  }
+}
+```
+
+## Link to slide:
+
+https://snappify.com/view/888e678f-01de-43fd-a79c-e92ff9a07d97
